more modification
git-svn-id: http://fsnet.googlecode.com/svn/trunk@492 07c314b0-eee8-11dd-b478-27b5dcd53177

+++ fr.univartois.ili.fsnet/fsnetWeb/src/main/java/fr/univartois/ili/fsnet/actions/ManageInterests.java
@@ -18,13 +18,8 @@ package fr.univartois.ili.fsnet.actions;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Date;
-import java.util.Iterator;
 import java.util.List;
-import java.util.ListIterator;
 import java.util.logging.Logger;
-
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
@@ -32,8 +27,6 @@ import javax.persistence.RollbackException;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import org.apache.commons.beanutils.BeanUtils;
 import org.apache.struts.action.ActionErrors;
 import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
@@ -41,10 +34,9 @@ import org.apache.struts.action.ActionMapping;
 import org.apache.struts.action.ActionMessage;
 import org.apache.struts.action.DynaActionForm;
 import org.apache.struts.actions.MappingDispatchAction;
-
 import fr.univartois.ili.fsnet.entities.EntiteSociale;
 import fr.univartois.ili.fsnet.entities.Interet;
-import fr.univartois.ili.fsnet.entities.Manifestation;
+
 
 /**
  * Execute CRUD Actions (and more) for the entity interet
@@ -88,7 +80,7 @@ public class ManageInterests extends MappingDispatchAction implements CrudAction
     public ActionForward modify(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
     	EntityManager em = factory.createEntityManager();
 		DynaActionForm dynaForm = (DynaActionForm) form;
-		String interestId = (String) dynaForm.get("interestId");
+		int interestId = (Integer)dynaForm.get("interestId");
 		String interestName = (String) dynaForm.get("interestName");
 				
         logger.info("interest modification: " + interestName);
@@ -116,26 +108,28 @@ public class ManageInterests extends MappingDispatchAction implements CrudAction
     @Override
     public ActionForward delete(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
     	EntityManager em = factory.createEntityManager();
-		DynaActionForm dynaForm = (DynaActionForm) form;
-		String interestId = (String) dynaForm.get("interestId");
 				
-        logger.info("interest deleted: id=" + interestId);
-        
-        try {
-            em.getTransaction().begin();
-            em.createQuery("DELETE FROM Interet WHERE idInteret = ?interestId")
-            	.setParameter("interestId", interestId)
-            	.executeUpdate();
-            em.getTransaction().commit();
-        } catch (RollbackException ex) {
-            ActionErrors actionErrors = new ActionErrors();
-            ActionMessage msg = new ActionMessage(
-                    "interest.notExists");
-            actionErrors.add("interest", msg);
-            saveErrors(request, actionErrors);
-        }
-
-        em.close();
+		if (request.getParameterMap().containsKey("interestId")) {
+			int interestId = Integer.valueOf(request.getParameter("interestId"));
+						
+	        logger.info("interest deleted: id=" + interestId);
+	        
+	        try {
+	            em.getTransaction().begin();
+	            em.createQuery("DELETE FROM Interet WHERE idInteret = ?interestId")
+	            	.setParameter("interestId", interestId)
+	            	.executeUpdate();
+	            em.getTransaction().commit();
+	        } catch (RollbackException ex) {
+	            ActionErrors actionErrors = new ActionErrors();
+	            ActionMessage msg = new ActionMessage(
+	                    "interest.notExists");
+	            actionErrors.add("interest", msg);
+	            saveErrors(request, actionErrors);
+	        }
+	
+	        em.close();
+		}
 		
 		return mapping.findForward("success");
     }
@@ -148,13 +142,31 @@ public class ManageInterests extends MappingDispatchAction implements CrudAction
 				
         logger.info("search interest: " + interestName);
         
-        // TO DO em.find
+        List<Interet> result =
+        	em.createQuery("SELECT interest FROM Interest interest WHERE interest.nomInteret LIKE '%:interestName%' ")
+        		.setParameter("interestName", interestName)
+        		.getResultList();
+        
+        request.getSession().setAttribute("interestResult", result);
 		
 		return mapping.findForward("success");
     }
 
     @Override
     public ActionForward display(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
-        throw new UnsupportedOperationException("Not supported yet.");
+    	/*EntityManager em = factory.createEntityManager();
+		DynaActionForm dynaForm = (DynaActionForm) form;
+		String interestId = dynaForm.get("interestId");
+				
+        logger.info("search interest: " + interestName);
+        
+        List<Interet> result =
+        	em.createQuery("SELECT interest FROM Interest interest WHERE interest.nomInteret LIKE '%:interestName%' ")
+        		.setParameter("interestName", interestName)
+        		.getResultList();
+        
+        request.getSession().setAttribute("interestResult", result);
+		*/
+		return mapping.findForward("success");
     }
 }
